Commodity Price Predictor/predict-n-sell/requirements.txt
streamlit
sklearn
pillow